move disenchantment logic to anvil mechanic
